chore: consolidate schema and component dictionaries
