+++ sites/fast-component-explorer/package.json
@@ -82,6 +82,7 @@
     "@microsoft/fast-components-foundation-react": "^3.1.8",
     "@microsoft/fast-jss-utilities": "^4.7.12",
     "@microsoft/fast-layouts-react": "^4.11.9",
+    "@microsoft/site-utilities": "^0.0.1",
     "@types/jest": "^25.2.1",
     "@types/lodash-es": "^4.17.0",
     "@types/node": "^10.12.12",
@@ -129,8 +130,8 @@
     "@microsoft/fast-components-react-msft": "^4.30.10",
     "@microsoft/fast-components-styles-msft": "^4.28.9",
     "@microsoft/fast-jss-manager-react": "^4.7.2",
-    "@microsoft/fast-tooling-react": "^2.0.3",
-    "@microsoft/fast-tooling": "^0.3.0",
+    "@microsoft/fast-tooling-react": "^2.0.4",
+    "@microsoft/fast-tooling": "^0.3.1",
     "@microsoft/fast-web-utilities": "^4.4.5",
     "react-dnd-html5-backend": "^9.0.0",
     "react-sortable-hoc": "^1.9.1"
