+++ sites/fast-creator/app/msft-components/README.md
@@ -3,6 +3,4 @@
 This folder contains:
 
 1. The unpublished schemas that correspond with the tooling v2 update
-2. A dictionary of msft react components - `dictionary.ts`
-3. Example data to include with the components when they are created - `example-data.ts`
-4. Wrapped msft react components - `wrapped-components.ts`
+2. Example data to include with the components when they are created - `example-data.ts`
