+++ sites/fast-creator/package.json
@@ -42,6 +42,7 @@
     "@microsoft/fast-jss-manager-react": "^4.7.2",
     "@microsoft/fast-jss-utilities": "^4.7.12",
     "@microsoft/fast-layouts-react": "^4.11.9",
+    "@microsoft/site-utilities": "^0.0.1",
     "@types/jest": "^25.2.1",
     "@types/node": "^9.6.7",
     "@types/react": "^16.8.0",
