+++ sites/fast-creator/webpack.config.js
@@ -98,7 +98,7 @@ module.exports = (env, args) => {
                     "react-dnd-html5-backend"
                 ),
                 react: path.resolve(rootNodeModules, "react"),
-                "react-dnd": path.resolve(nodeModules, "react-dnd"),
+                "react-dnd": path.resolve(rootNodeModules, "react-dnd"),
                 "react-dom": path.resolve(rootNodeModules, "react-dom"),
             },
         },
